[SYCL] Infer IdT in GroupBroadcast from group type
- Work-group => any integral type
- Sub-group => unsigned 32-bit integer
